Add Mac networking/socket implementation.
1. Repurpose linux/socket.cc for all POSIX by making a few trivial
changes to use the common POSIX/BSD socket APIs.

2. Add Mac impl of network_interface.cc (to discover network interfaces).
Introduced a platform::InterfaceIndex type to replace the use of int32_t
everywhere.

3. Add std::ostream operators for IPAddress/IPEndpoint, and the
network_interface.h classes (logging convenience), and updated the demo
apps to print out detected network interfaces.

+++ api/impl/testing/fake_mdns_platform_service.h
@@ -22,7 +22,7 @@ class FakeMdnsPlatformService final : public MdnsPlatformService {
 
   // PlatformService overrides.
   std::vector<BoundInterface> RegisterInterfaces(
-      const std::vector<int32_t>& interface_index_whitelist) override;
+      const std::vector<int64_t>& interface_index_whitelist) override;
   void DeregisterInterfaces(
       const std::vector<BoundInterface>& registered_interfaces) override;
